First round of fixes. (#13452)
### Description
First round of fixes for C4244 error.



### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->

+++ onnxruntime/contrib_ops/cpu/activations.h
@@ -82,7 +82,7 @@ class Gelu : public OpKernel {
             p_output[i] = value * static_cast<T>(M_SQRT1_2);
           }
 
-          MlasComputeErf(p_output, p_output, count);
+          MlasComputeErf(p_output, p_output, gsl::narrow_cast<size_t>(count));
 
           for (int64_t i = 0; i < count; i++) {
             p_output[i] = 0.5f * p_input[i] * (p_output[i] + 1.0f);
